pyproject.toml
[project]
name = "xclaw"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    # GPU 推理（从 PyTorch 专用源安装）
    "torch>=2.5.0",
    "torchvision>=0.20.0",

    # OmniParser 依赖
    "ultralytics>=8.0.0",
    "transformers>=4.40.0,<4.46.0",
    "safetensors>=0.4.0",
    "accelerate>=0.30.0",
    "einops>=0.8.0",
    "timm>=1.0.0",
    "supervision>=0.25.0",
    "easyocr>=1.7.0",
    "paddlepaddle>=2.6.0",
    "paddleocr>=2.7.0",
    "opencv-python-headless>=4.8.0",
    "openai>=1.3.0",

    # 截图
    "mss>=9.0.0",

    # 鼠标键盘控制
    "pyautogui>=0.9.54",

    # 图像处理
    "Pillow>=10.0.0",

    # CLI
    "click>=8.0.0",
    "rich>=13.0.0",

    # 模型下载
    "huggingface-hub[cli]>=0.25.0",
]

[tool.uv]
package = true

# 关键改动：只让 torch 和 torchvision 走 PyTorch 源，其他包走默认 PyPI
[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu121" }]
torchvision = [{ index = "pytorch-cu121" }]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[project.scripts]
xclaw = "xclaw.cli:main"
